requirement.txt
absl-py==0.15.0
antlr4-python3-runtime==4.9.3
asgiref==3.5.2
astunparse==1.6.3
attrs==21.4.0
autopep8==1.6.0
black==22.3.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.5
cycler==0.11.0
Cython==0.29.30
dataclasses==0.6
Django==3.2
djangorestframework==3.13.1
fire==0.4.0
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.3.3
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hydra-core==1.2.0
idna==3.3
imageio==2.19.3
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==4.11.4
importlib-resources==5.8.0
iopath==0.1.9
jsonschema==4.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
labelme2coco==0.2.1
lxml==4.9.0
Markdown==3.3.7
matplotlib==3.5.2
mypy-extensions==0.4.3
networkx==2.8.4
numpy==1.19.5
oauthlib==3.2.0
omegaconf==2.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
pixellib==0.7.1
platformdirs==2.5.2
portalocker==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydot==1.4.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pywin32==304
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
sahi==0.9.4
scikit-image==0.19.3
scipy==1.8.1
Shapely==1.8.2
six==1.15.0
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminaltables==3.1.10
tifffile==2022.5.4
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.12.1
xmljson==0.2.1
yacs==0.1.8
zipp==3.8.0
